+++ demo/app/index.js
@@ -11,6 +11,8 @@ const demos = {
   treeGroups: require('./demo-tree-groups').default,
   linkedTimelines: require('./demo-linked-timelines').default,
   elementResize: require('./demo-element-resize').default,
+  stickyHeader: require('./demo-sticky-header').default,
+  customItems: require('./demo-cusom-items').default,
   stickyHeader: require('./demo-sticky-header').default
 }