+++ src/index.js
@@ -1,3 +1,10 @@
 import Timeline from './lib/Timeline'
 
+export {
+  default as TimelineMarkers
+} from './lib/markers/public/TimelineMarkers'
+export { default as TodayMarker } from './lib/markers/public/TodayMarker'
+export { default as CustomMarker } from './lib/markers/public/CustomMarker'
+export { default as CursorMarker } from './lib/markers/public/CursorMarker'
+
 export default Timeline