+++ src/lib/layout/Header.js
@@ -31,7 +31,19 @@ class Header extends Component {
       width,
       stickyOffset,
       stickyHeader,
-      headerRef
+      headerRef,
+      hasRightSidebar,
+      showPeriod,
+      canvasTimeStart,
+      canvasTimeEnd,
+      canvasWidth,
+      minUnit,
+      timeSteps,
+      headerLabelFormats,
+      subHeaderLabelFormats,
+      headerLabelGroupHeight,
+      headerLabelHeight,
+      registerScroll
     } = this.props
 
     const headerStyle = {
@@ -43,18 +55,27 @@ class Header extends Component {
     return (
       <div
         className={`rct-header-container ${headerClass}`}
-        data-test-id="timeline-elements-container"
+        data-testid="timeline-elements-container"
         ref={headerRef}
         style={headerStyle}
       >
         {leftSidebarHeader}
-        <div
-          style={{ width }}
-          data-test-id="timeline-elements-header-container"
-        >
+        <div style={{ width }} data-testid="timeline-elements-header-container">
           <TimelineElementsHeader
-            data-test-id="timeline-elements-header"
-            {...this.props}
+            data-testid="timeline-elements-header"
+            hasRightSidebar={hasRightSidebar}
+            showPeriod={showPeriod}
+            canvasTimeStart={canvasTimeStart}
+            canvasTimeEnd={canvasTimeEnd}
+            canvasWidth={canvasWidth}
+            minUnit={minUnit}
+            timeSteps={timeSteps}
+            width={width}
+            headerLabelFormats={headerLabelFormats}
+            subHeaderLabelFormats={subHeaderLabelFormats}
+            headerLabelGroupHeight={headerLabelGroupHeight}
+            headerLabelHeight={headerLabelHeight}
+            registerScroll={registerScroll}
           />
         </div>
         {rightSidebarHeader}