+++ src/lib/row/GroupRow.js
@@ -6,6 +6,7 @@ class GroupRow extends Component {
   static propTypes = {
     onClick: PropTypes.func.isRequired,
     onDoubleClick: PropTypes.func.isRequired,
+    onContextMenu: PropTypes.func.isRequired,
     isEvenRow: PropTypes.bool.isRequired,
     style: PropTypes.object.isRequired,
     clickTolerance: PropTypes.number.isRequired,
@@ -15,6 +16,7 @@ class GroupRow extends Component {
 
   render() {
     const {
+      onContextMenu,
       onDoubleClick,
       isEvenRow,
       style,
@@ -32,6 +34,7 @@ class GroupRow extends Component {
     return (
       <PreventClickOnDrag clickTolerance={clickTolerance} onClick={onClick}>
         <div
+          onContextMenu={onContextMenu}
           onDoubleClick={onDoubleClick}
           className={(isEvenRow ? 'rct-hl-even ' : 'rct-hl-odd ') + (classNamesForGroup ? classNamesForGroup.join(' ') : '')}
           style={style}