+++ CHANGELOG.md
@@ -7,10 +7,64 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ## Unreleased
 
+
 ### Added
 
 * possibility to add custom classes to the group-rows
 * possibility to add custom height to row (tree-group example)
+* rework item renderer to render the whole item using render prop and prop getters
+
+### Breaking
+* add scroll ref
+
+### 0.18.2
+
+
+
+### Fixed
+
+* `onCanvasClick` not fired - #383
+* cursor marker disappear while hovering over item - #378
+
+### 0.18.1
+
+### Fixed
+
+* Date passed to CursorMarker child is wrong - #379
+* groupRenderer doesnt work for right sidebar - #377
+
+### 0.18.0
+
+### Fixed
+
+* Timeline now respects changes to `headerLabelFormats` and `subHeaderLabelFormats` - #362
+
+### Added
+
+* timeline markers - user can have more control over markers that are rendered on the timeline. See `TimelineMarkers` section of README for documentation - #327
+
+### Breaking
+
+* Removed support for React 15 and lower.  This is due to the fact that 16+ supports returning arrays from render, something that the TimelineMarker feature relies on.
+* removed `showCursorLine` prop in favor of using the `CursorMarker` component. See `TimelineMarkers` section of README for documentation.
+
+```diff
+import Timeline,
++ {TimelineMarkers, CursorMarker}
+from 'react-calendar-timeline'
+
+<Timeline
+- showCursorLine
+
+- />
++ >
+
++ <TimelineMarkers>
++  <CursorMarker />
++ </TimelineMarkers>
+
++ </Timeline>
+```
 
 ### 0.17.3